+++ src/main/java/kr/codeit/relaxtogether/repository/review/ReviewRepository.java
@@ -1,7 +1,8 @@
-package kr.codeit.relaxtogether.repository;
+package kr.codeit.relaxtogether.repository.review;
 
 import kr.codeit.relaxtogether.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
+
 }
