+++ src/test/java/kr/codeit/relaxtogether/service/ReviewServiceTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.Assertions.tuple;
 import kr.codeit.relaxtogether.dto.review.request.WriteReviewRequest;
 import kr.codeit.relaxtogether.entity.User;
 import kr.codeit.relaxtogether.entity.gathering.Gathering;
-import kr.codeit.relaxtogether.repository.ReviewRepository;
 import kr.codeit.relaxtogether.repository.UserRepository;
 import kr.codeit.relaxtogether.repository.gathering.GatheringRepository;
+import kr.codeit.relaxtogether.repository.review.ReviewRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
